extract server seed logic from index.js to seedManager.js for better testability.
